fix(linalg): preserve FP64 Jacobi coefficients

+++ src/cfdx/core/linalg/preconditioner.h
@@ -59,12 +59,12 @@ public:
     bool setup(const SparseMatrix& A) override {
         if (A.n_rows()!=A.n_cols()) return false;
         const auto* row=A.row_offsets_data(); const auto* col=A.columns_data(); const auto* val=A.values_data();
-        inv_diag_.assign(A.n_rows(),0.0f);
+        inv_diag_.assign(A.n_rows(),0.0);
         for(std::size_t i=0;i<A.n_rows();++i){
             bool found=false;
             for(std::size_t k=row[i];k<row[i+1];++k) if(col[k]==i){
                 if(!std::isfinite(val[k]) || val[k]==0.0) { inv_diag_.clear(); return false; }
-                inv_diag_[i]=static_cast<float>(1.0/val[k]); found=true; break;
+                inv_diag_[i]=1.0/val[k]; found=true; break;
             }
             if(!found) { inv_diag_.clear(); return false; }
         }
@@ -74,7 +74,7 @@ public:
         if(r.size()!=inv_diag_.size() || z.size()!=r.size()) return false;
         for(std::size_t i=0;i<r.size();++i){
             if(precision_==SolverPrecision::FP32)
-                z(i)=static_cast<double>(static_cast<float>(r(i))*inv_diag_[i]);
+                z(i)=static_cast<double>(static_cast<float>(r(i))*static_cast<float>(inv_diag_[i]));
             else
                 z(i)=static_cast<double>(r(i))*static_cast<double>(inv_diag_[i]);
         }
@@ -83,7 +83,7 @@ public:
     const char* name() const override { return "MixedPrecisionJacobi"; }
 private:
     SolverPrecision precision_;
-    std::vector<float> inv_diag_;
+    std::vector<double> inv_diag_;
 };
 
 } // namespace cfdx::core
